% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupar_isco.R
\name{isco08labels}
\alias{isco08labels}
\title{Get labels for ISCO-08}
\usage{
isco08labels(isco08, display.nas = FALSE)
}
\arguments{
\item{isco08}{a numeric vector with the ISCO-08 codes}

\item{display.nas}{boolean. If true (defaul is \code{FALSE}), displays table with unmatched ISCO-08 codes.}
}
\value{
It returns a vector with ISCO-08 labels.
}
\description{
The function gets labels for ISCO-08.
}
