% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupar_isco.R
\name{isco08to88}
\alias{isco08to88}
\title{Recode ISCO-08 into ISCO-88}
\usage{
isco08to88(isco08, display.nas = FALSE)
}
\arguments{
\item{isco08}{a numeric vector with the ISCO-08 codes}

\item{display.nas}{boolean. If true (defaul is \code{FALSE}), displays table with unmatched ISCO-08 codes.}
}
\value{
It returns a vector with ISCO-88 codes.
}
\description{
The function recodes ISCO-08 into ISCO-88.
}
