% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupar_occupation.R
\name{isco08toISEI92}
\alias{isco08toISEI92}
\title{Recode ISCO-08 into ISEI (a.k.a. ISEI-92)}
\usage{
isco08toISEI92(isco08, display.nas = FALSE)
}
\arguments{
\item{display.nas}{boolean. If true (defaul is \code{FALSE}), displays table with unmatched ISCO-08 codes.}

\item{isco88}{a numeric vector with the ISCO-08 codes}
}
\value{
It returns a vector with ISEI-92 codes.
}
\description{
The function recodes ISCO-08 into ISEI-92.
}
\details{
Ganzeboom, H. B., De Graaf, P. M., & Treiman, D. J.,  (1992) A standard international socio-economic index of occupational status, Social science research, 21(1), 1–56.
}
