% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupar_isco.R
\name{isco68to88}
\alias{isco68to88}
\title{Recode ISCO-68 into ISCO-88}
\usage{
isco68to88(isco68, display.nas = FALSE)
}
\arguments{
\item{isco68}{a numeric vector with the ISCO-68 codes}

\item{display.nas}{boolean. If true (defaul is \code{FALSE}), displays table with unmatched ISCO-68 codes.}
}
\value{
It returns a vector with ISCO-88 codes.
}
\description{
The function recodes ISCO-68 into ISCO-88.
}
