% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupar_isco.R
\name{isco88to08}
\alias{isco88to08}
\title{Recode ISCO-88 into ISCO-08}
\usage{
isco88to08(isco88, display.nas = FALSE)
}
\arguments{
\item{isco88}{a numeric vector with the ISCO-88 codes}

\item{display.nas}{boolean. If true (defaul is \code{FALSE}), displays table with unmatched ISCO-88 codes.}
}
\value{
It returns a vector with ISCO-08 codes.
}
\description{
The function recodes ISCO-88 into ISCO-08.
}
