% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupar_occupation.R
\name{isco88toEGP}
\alias{isco88toEGP}
\title{Recode occupation to EGP class scheme}
\usage{
isco88toEGP(
  isco88,
  n.employees = NULL,
  self.employed = NULL,
  collapse = FALSE,
  n.classes = 11
)
}
\arguments{
\item{isco88}{numeric vector with the isco88 4 digit code.}

\item{n.employees}{a vector containing the number of employees.}

\item{self.employed}{a vector containing the value 1 whenever the person is self-employed}

\item{collapse}{deprecated}

\item{n.classes}{either 11, 7, 5, or 3, indicating the number of EGP classes to use}
}
\description{
The function creates a factor with the EGP class scheme
}
\details{
The number of classes to use in the EGP scheme is based on
- Connelly, R., Gayle, V., & Lambert, P. S., (2016) A review of occupation-based social classifications for social survey research, Methodological Innovations, v.9, p.1-14.
}
