% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/occupar_occupation.R
\name{isco88toISEI92}
\alias{isco88toISEI92}
\title{Recode ISCO-88 into ISEI (a.k.a. ISEI-92)}
\usage{
isco88toISEI92(isco88, display.nas = FALSE)
}
\arguments{
\item{isco88}{a numeric vector with the ISCO-88 codes}

\item{display.nas}{boolean. If true (defaul is \code{FALSE}), displays table with unmatched ISCO-88 codes.}
}
\value{
It returns a vector with ISEI-92 codes.
}
\description{
The function recodes ISCO-88 into ISEI-92.
}
\details{
Ganzeboom, H. B., De Graaf, P. M., & Treiman, D. J.,  (1992) A standard international socio-economic index of occupational status, Social science research, 21(1), 1–56.
}
